+++ fabric/src/main/java/com/gitlab/srcmc/rctmod/fabric/FabricCommon.java
@@ -21,6 +21,7 @@ import com.gitlab.srcmc.rctmod.ModCommon;
 import com.gitlab.srcmc.rctmod.api.RCTMod;
 
 import fuzs.forgeconfigapiport.fabric.api.neoforge.v4.NeoForgeConfigRegistry;
+import fuzs.forgeconfigapiport.fabric.api.neoforge.v4.NeoForgeModConfigEvents;
 import net.fabricmc.api.ModInitializer;
 import net.neoforged.fml.config.ModConfig;
 
@@ -29,5 +30,13 @@ public class FabricCommon implements ModInitializer {
     public void onInitialize() {
         ModCommon.init();
         NeoForgeConfigRegistry.INSTANCE.register(ModCommon.MOD_ID, ModConfig.Type.SERVER, RCTMod.getInstance().getServerConfig().getSpec());
+        NeoForgeModConfigEvents.loading(ModCommon.MOD_ID).register(FabricCommon::onConfigLoadingOrReloading);
+        NeoForgeModConfigEvents.reloading(ModCommon.MOD_ID).register(FabricCommon::onConfigLoadingOrReloading);
+    }
+
+    static void onConfigLoadingOrReloading(ModConfig config) {
+        if(config.getType() == ModConfig.Type.SERVER) {
+            RCTMod.getInstance().getServerConfig().reload();
+        }
     }
 }